requirements.txt
google-generativeai==0.8.3
Pillow==10.4.0
python-dotenv==1.1.1
requests==2.32.4
Flask==3.0.0
Werkzeug==3.0.1
